typescript(vx-demo): rewrite Text example in TypeScript
